refactor(benchmarks): extract conformance surfaces
